feat: created FormInfoContentConainter
